requirements.txt
langgraph~=0.6.3
openai~=1.99.1
python-dotenv~=1.1.1

requests~=2.32.4
httpx~=0.28.1
pydantic~=2.11.7
Flask~=3.1.1